rename color file
