+++ packages/cli/src/colors/utils.ts
